Refactor tree drawing code to tree emitter
